requirements.txt
bcrypt==4.2.1
blinker==1.9.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==3.1.0
Flask-JWT-Extended==4.7.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
PyJWT==2.10.1
requests==2.32.3
urllib3==2.2.3
Werkzeug==3.1.3
